FI-2342: restore app.tsx, move events to separate file, update messaging to completed for each action to get handled more fluidly in LinkUI, handle type issues when creating config.  Update to latest wagmi and wagmi/core to get connectors that can handle injected().
