+++ packages/link/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@meshconnect/web-link-sdk",
-  "version": "2.0.6",
+  "version": "2.1.0",
   "description": "A client-side JS library for integrating with Mesh Connect",
   "main": "./src/index.ts",
   "module": "./src/index.ts",
@@ -10,24 +10,24 @@
   "repository": "https://github.com/FrontFin/mesh-web-sdk.git",
   "homepage": "https://github.com/FrontFin/mesh-web-sdk",
   "devDependencies": {
+    "@types/jest": "^29.5.12",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
     "eslint": "^7.23.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^3.4.0",
-    "prettier": "^2.2.1",
-    "typescript": "^4.6.4",
-    "ts-jest": "^29.1.2",
     "jest": "^29.7.0",
-    "@types/jest": "^29.5.12",
-    "jest-environment-jsdom": "^29.7.0"
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "^2.2.1",
+    "ts-jest": "^29.1.2",
+    "typescript": "^4.6.4"
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.17.9",
-    "@wagmi/core": "^2.6.16",
-    "viem": "^2.x",
-    "wagmi": "^2.5.7"
+    "@meshconnect/node-api": "^2.0.5",
+    "@wagmi/connectors": "^5.0.7",
+    "@wagmi/core": "^2.10.5",
+    "viem": "^2.x"
   },
   "scripts": {
     "build": "yarn updateVersion && rimraf dist && yarn build:esm && yarn build:cjs && yarn copy",
