+++ packages/link/src/utils/event-types.ts
@@ -5,8 +5,6 @@ export type LinkEventType =
   | IntegrationConnectionError
   | TransferCompleted
   | IntegrationSelected
-  | IntegrationInjectedWalletSelected
-  | IntegrationChainSwitchRequest
   | CredentialsEntered
   | TransferStarted
   | TransferPreviewed
@@ -20,8 +18,6 @@ export type LinkEventType =
   | TransferAssetSelected
   | TransferNetworkSelected
   | TransferAmountEntered
-  | TransferBalanceRequest
-  | TransferInjectedRequest
   | TransferMfaRequired
   | TransferMfaEntered
   | TransferKycRequired
@@ -37,8 +33,6 @@ const LINK_EVENT_TYPE_KEYS = [
   'integrationAccountSelectionRequired',
   'transferCompleted',
   'integrationSelected',
-  'integrationInjectedWalletSelected',
-  'integrationChainSwitchRequest',
   'credentialsEntered',
   'transferStarted',
   'transferPreviewed',
@@ -46,8 +40,6 @@ const LINK_EVENT_TYPE_KEYS = [
   'transferExecutionError',
   'pageLoaded',
   'transferAssetSelected',
-  'transferBalanceRequest',
-  'transferInjectedRequest',
   'transferNetworkSelected',
   'transferAmountEntered',
   'transferMfaRequired',
@@ -86,39 +78,6 @@ export interface IntegrationConnectionError extends LinkEventBase {
 export interface TransferCompleted extends LinkEventBase {
   type: 'transferCompleted'
   payload: TransferFinishedPayload
-}
-export interface IntegrationChainSwitchRequest extends LinkEventBase {
-  type: 'integrationChainSwitchRequest'
-  payload: {
-    chainId: number
-  }
-}
-
-export interface TransferInjectedRequest extends LinkEventBase {
-  type: 'transferInjectedRequest'
-  payload: {
-    amount: number
-    toAddress: string
-    decimalPlaces: number
-    chainId: number
-    account: string
-  }
-}
-
-export interface IntegrationInjectedWalletSelected extends LinkEventBase {
-  type: 'integrationInjectedWalletSelected'
-  payload: {
-    integrationType: string
-    integrationName: string
-  }
-}
-
-export interface TransferBalanceRequest extends LinkEventBase {
-  type: 'transferBalanceRequest'
-  payload: {
-    account: string
-    chainId: number
-  }
 }
 
 export interface IntegrationSelected extends LinkEventBase {
@@ -264,107 +223,4 @@ export interface SessionSymmary {
     networkId?: string
   }
   errorMessage?: string
-}
-
-export type WalletBrowserEventType =
-  | WalletBrowserInjectedWalletSelected
-  | WalletBrowserChainSwitchRequest
-  | WalletBrowserTransferBalanceRequest
-  | WalletBrowserNativeTransferRequest
-  | WalletBrowserNonNativeTransferRequest
-  | WalletBrowserNativeSmartDeposit
-  | WalletBrowserNonNativeSmartDeposit
-  | WalletBrowserDisconnect
-
-const WALLET_BROWSER_EVENT_TYPE_KEYS = [
-  'walletBrowserInjectedWalletSelected',
-  'walletBrowserChainSwitchRequest',
-  'walletBrowserTransferBalanceRequest',
-  'walletBrowserNativeTransferRequest',
-  'walletBrowserNonNativeTransferRequest',
-  'walletBrowserNativeSmartDeposit',
-  'walletBrowserNonNativeSmartDeposit',
-  'walletBrowserDisconnect'
-] as const
-
-export type WalletBrowserEventTypeKeys =
-  (typeof WALLET_BROWSER_EVENT_TYPE_KEYS)[number]
-
-export function isWalletBrowserEventTypeKey(
-  key: string
-): key is WalletBrowserEventTypeKeys {
-  return WALLET_BROWSER_EVENT_TYPE_KEYS.includes(
-    key as WalletBrowserEventTypeKeys
-  )
-}
-
-interface WalletBrowserEventBase {
-  type: WalletBrowserEventTypeKeys
-}
-
-export interface WalletBrowserInjectedWalletSelected
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserInjectedWalletSelected'
-  payload: {
-    integrationName: string
-  }
-}
-
-export interface WalletBrowserChainSwitchRequest
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserChainSwitchRequest'
-  payload: {
-    chainId: number
-  }
-}
-
-export interface WalletBrowserTransferBalanceRequest
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserTransferBalanceRequest'
-  payload: {
-    account: string
-    chainId: number
-  }
-}
-
-export interface WalletBrowserNativeTransferRequest
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserNativeTransferRequest'
-  payload: {
-    toAddress: string
-    amount: number
-    decimalPlaces: number
-    chainId: number
-    account: string
-  }
-}
-
-interface SmartContractPayload {
-  address: string
-  abi: string
-  functionName: string
-  args: any[]
-  value?: bigint
-}
-
-export interface WalletBrowserNonNativeTransferRequest
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserNonNativeTransferRequest'
-  payload: SmartContractPayload
-}
-
-export interface WalletBrowserNativeSmartDeposit
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserNativeSmartDeposit'
-  payload: SmartContractPayload
-}
-
-export interface WalletBrowserNonNativeSmartDeposit
-  extends WalletBrowserEventBase {
-  type: 'walletBrowserNonNativeSmartDeposit'
-  payload: SmartContractPayload
-}
-
-export interface WalletBrowserDisconnect extends WalletBrowserEventBase {
-  type: 'walletBrowserDisconnect'
 }
